requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.23.0,<1.0.0
rembg>=2.0.56
Pillow>=10.0.0
python-multipart>=0.0.9
typer>=0.9.0
pytest>=7.4.0
onnxruntime>=1.16.0
